Fix token field type from int64 to int32

+++ BRM/source/sys/fm_nai3_pol/fm_nai_act_rate.c
@@ -148,8 +148,8 @@ op_nai_act_rate(
         int32 *model_code = (int32 *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_MODELTYPE_NAI3, 0, ebufp);
         pin_decimal_t *trans_id = (pin_decimal_t *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_TRANSACTION_ID, 0, ebufp);
         char * prompt_txt = (char *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_PROMPT_NAI3, 0, ebufp);
-        int32 * tokens_in = (int64 *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_INPUT_TOKENS_NAI3, 0, ebufp);
-        int32 * tokens_out = (int64 *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_OUTPUT_TOKENS_NAI3, 0, ebufp);
+        int32 * tokens_in = (int32 *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_INPUT_TOKENS_NAI3, 0, ebufp);
+        int32 * tokens_out = (int32 *)PIN_FLIST_FLD_GET(i_flistp, PIN_FLD_OUTPUT_TOKENS_NAI3, 0, ebufp);
 
         /*
          * service_flist will hold the PIN_FLD_RESULTS
